Adding toDescription() function for enum objects.
This generates a simple description for an enum value by adding spaces between the words forming an enum field. For example, if an enum field is ActiveParticiple, the generated description is "Active Participle". Notice that this assumes the enum fields to be upper camel cased, i.e. each word (including the first word) starts with a capital letter.